update price feed readme
In the first draft of the price-feed server I introduced the FNV1A64 hash
of an asset symbol as it's ID.
This turned out to be a pita when adding a new asset to the oracle.

We moved from the FNV1A64 hash to an arbitrary index defined by the feed server.