use streams to represent feeds
I moved from the previous implementation to a stream based one.
Feeds are represented as a tuple of both a stream of notification and a joinable
handle.
Feeds should take the `keep_running` and terminate whenever the value move from
true to false.

The choice of using `start` instead of `new` is purely based on clippy, as `new`
is usually used to return `Self`.