remove ignore_min_transfer from vesting
I initially added `ignore_min_transfer` because of the first bonded finance
pallet implementation.

This bool was allowing us to do vested transfer <= MinVestedTransfer from the
Vesting pallet because any user was able to bond any amount. Hence resulting in
a reward share being possibly < MinVestedTransfer.

This is not the case anymore with the parts * price model.