forbid impossible state by avoiding default instance for RelayerConfig
We were using Default + ValueQuery for RelayerConfig.
It was fine because RelayerConfig was containing an `Option<AccountId>`
but this would allow inconsistent states such as None for RelayerConfig.current
& None for RelayerConfig.next.
By removing the option within RelayerConfig and using an OptionQuery for the
storage,
we it's more clear that a Relayer can be Set/Unset. And if set, an AcccountId
must be present.