feat(tests): custom precision for proptests to avoid edge cases (#702)
* feat(tests): custom precision for proptests to avoid edge cases

The issue with proptest is that because of the fact that we generate random
values, we might have some issue because of division. The issue rely on the fact
that for instane, integer division might result in less/more value than expected (odd number
for instance). We can circumvent thoses cases (pretty rare but still making the
CI fail) by lowering the precision of the expected value for some tests.

* fix(style): apply make style

* fix(lint): avoid identity conversion