[CU-206yhnr] ComposableJS (#1198)
## Issue
https://app.clickup.com/t/206yhnr

## Description
~~*Please replace this line with info that is not in the ClickUp ticket, eg: "also bumps dependency foo version"*~~

I've moved the oracle setup to our new ComposableJS package, cause i needed a lab rat for testing everything. :D 

Some intended checkpoints like:
- All missing types added
- CI Setup
will be skipped for now, and will be worked on during this sprint. We want this package to get merged so we can get it into the frontend packages as quickly as possible.



Co-authored-by: Dominik Roth <96540347+Dom-Roth@users.noreply.github.com>
Co-authored-by: 0xbrainjar <83704623+0xBrainjar2@users.noreply.github.com>