Initial refactor to make things ready for transfer
Transfer from/to Picasso<> Kusama
Transfer from/to Picasso <> Karura

KUSD is used instead of aUSD, since aUSD is the same token, with a different name on different networks.

Balances are now fetchable across all parachains, with the store being updated.

Transaction fee is automatic in this PR and it is taken from the initial asset, requires a new set of API calls to fetch that value from other sources.