migrated lfs tracked files (so these files stored in blob instead of diffed) (#2707)
Signed-off-by: Dzmitry Lahoda <dzmitry@lahoda.pro>

## Issue

LFS was added and file extensions to LFS, but these file not migrated. 

After https://github.com/ComposableFi/composable/pull/2646 it appeared
that need https://github.com/ComposableFi/composable/pull/2694 and also
better to migrate (idea before was not to force migrate, but ensure all
new items in LFS). But seems migration is less noisy.

So this PR is result of running just one command to do migration
https://github.com/git-lfs/git-lfs/blob/main/docs/man/git-lfs-migrate.adoc?utm_source=gitlfs_site&utm_medium=doc_man_migrate_link&utm_campaign=gitlfs
. Nothing more.

Why? Because LFS has no history and may be avoided on some checkouts
which slows down monorepo degradation.

## Checklist

- @easteregg or other FE is ok for FE stuff
- @cor or @Martin-Kiss are ok for doc stuff
- @Dom-Roth 🥇 seems got `lfs detected` label violation first
https://github.com/ComposableFi/composable/pull/2630

- look at docasaurus after it released 
- look at UI

because most of files (only audit may be not) are from UI or docs

Signed-off-by: Dzmitry Lahoda <dzmitry@lahoda.pro>