PICA-136: upgrade to v0.9.33 (#2997)
#### Intro

Upgrade from v0.9.30 to v0.9.33

#### Migrations

- v0.9.31
  - [x] https://github.com/paritytech/polkadot/pull/6091
  - [x] https://github.com/paritytech/polkadot/pull/5729
    - [x] Preimage v1
    - [x] Scheduler v4
    - [x] Democracy v1
    - [x] Multisig v1
- v0.9.32
- v0.9.33

#### Notes

Are we forced to have the old migrations listed? Parity removed the v2
to v3 migration for the Scheduler, I changed it to a combination of v1
to v4 and v3 to v4 as they are idempotent.

I removed `price-feed` and associated packages as they are outdated and
not particularly useful.

The `local-integration-tests` package must be updated.