xcvm: simplify instructions handling in interpreter contract
XCVMProgram::instructions is a VecDeque.  The simplest way to split it
into first element and tail containing the rest of the instruction is
to use pop_front method.  In fact, this is a constant-time operation.

However, handle_execute_step does a strange dance where it converts
the instructions into an iterator and then collects the iterator into
a new VecDeque.  That’s an O(n) operation with an additional
allocation.

Simplify the code to use the pop_front instead.