xcvm: simplify instructions handling in interpreter contract (#3668)
XCVMProgram::instructions is a VecDeque. The simplest way to split it
into first element and tail containing the rest of the instruction is to
use pop_front method. In fact, this is a constant-time operation.

However, handle_execute_step does a strange dance where it converts the
instructions into an iterator and then collects the iterator into a new
VecDeque. That’s an O(n) operation with an additional allocation.

Simplify the code to use the pop_front instead.


- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [x] I have linked Zenhub/Github or any other reference item if one
exists
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [x] I waited and did best effort for `pr-workflow-check /
draft-release-check` to finish with success(green check mark) with my
changes
- [x] I have added at least one reviewer in reviewers list
- [x] I tagged(@) or used other form of notification of one person who I
think can handle best review of this PR
- [x] I have proved that PR has no general regressions of relevant
features and processes required to release into production