xcvm: avoid String allocations when serialising Displayed types
Firstly, use Serializer::collect_str to possibly avoid allocating
a string in case of serialisers which optimise that call.  In the
worst case, collect_str behaves the same way as old code.

Secondly, use Visitor with Visitor::visitor_str method implemented to
avoid having to allocate a String when deserialising the value.

Lastly, with all that, add some more documentation and tests for the
type.