xcvm: when testing read contracts from target directory directly
Change the test environment setup such that it reads contract files
from target directory directly rather than requiring environment
variables to be set pointing to the contracts.  This should ‘just
work’ in many situations so long as user builds the contracts.

To further help the user, introduce build-contracts.sh script which
builds contracts necessary by the test environment.

The remaining downside of the approach is that the contracts won’t be
automatically recompiled if user changes them and then executes tests
that use them.