xcvm: remove Network generic argument from xc_core::Instruction
It’s rather unlikely that network in an XCVM Instruction type will
ever be represented as something other than NetworkId.  Network is
part of the interface which needs to be understood by all the chains
in the same way and have the same representation.

Remove Network generic argument from xc_core::Instruction to simplify
the code slightly.