xcvm: remove Network generic argument from xc_core::Instruction (#3734)
It’s rather unlikely that network in an XCVM Instruction type will ever
be represented as something other than NetworkId. Network is part of the
interface which needs to be understood by all the chains in the same way
and have the same representation.

Remove Network generic argument from xc_core::Instruction to simplify
the code slightly.



- [ ] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [ ] I marked PR by `misc` label if it should not be in release notes
- [ ] I have linked Zenhub/Github or any other reference item if one
exists
- [ ] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [ ] I waited and did best effort for `pr-workflow-check /
draft-release-check` to finish with success(green check mark) with my
changes
- [ ] I have added at least one reviewer in reviewers list
- [ ] I tagged(@) or used other form of notification of one person who I
think can handle best review of this PR
- [ ] I have proved that PR has no general regressions of relevant
features and processes required to release into production