AR NetworkId (#3776)
- Introduced `NetworkId: u32 into Assets registry Pallet
- changed protocol_id to [u8; 4] from [u8.8]. Now protocol id are
expected to be Pallet::index() as u32. Pallet::index() is u64 but we
need 4 bytes. which is ok considering pallets' indices are low.
- Asset id for pallets is now concatenated 4 bytes of NetworkId, 4 bytes
of protocol_id, 8 bytes of nonce of a pallet
- added test to make sure bytes are concatenated correctly
- changed old tests
- updated LP tokens and old staking pallet tokens asset assignment to
use NetworkId and Pallet::index()

- [ ] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [ ] I marked PR by `misc` label if it should not be in release notes
- [ ] I have linked Zenhub/Github or any other reference item if one
exists
- [ ] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [ ] I waited and did best effort for `pr-workflow-check /
draft-release-check` to finish with success(green check mark) with my
changes
- [ ] I have added at least one reviewer in reviewers list
- [ ] I tagged(@) or used other form of notification of one person who I
think can handle best review of this PR
- [ ] I have proved that PR has no general regressions of relevant
features and processes required to release into production