pallet-cosmwasm: simplify Gas by using Vec rather than VecDeq (#3793)
Gas pushes and pops checkpoints only to one side of the checkpoints
field so there's no need for it to be a VecDeq.  A plain Vec suffices
and makes the compiled code simpler (VecDeq is a more complex data
structure than Vec).

- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [x] I have linked Zenhub/Github or any other reference item if one
exists
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [ ] I waited and did best effort for `pr-workflow-check /
draft-release-check` to finish with success(green check mark) with my
changes
- [ ] I have added at least one reviewer in reviewers list
- [ ] I tagged(@) or used other form of notification of one person who I
think can handle best review of this PR
- [ ] I have proved that PR has no general regressions of relevant
features and processes required to release into production