xcvm: change XCVMAck into an enum; and swap values (#3989)
XCVMAck can hold limited set of values (success or failure) so rather
than writing it as a newtype around u8 change it into an enum.  This
simplifies any code that needs to match on an XCVMAck value since it
doesn’t need to check values greater than 1.

Furthermore, rename the variants to Ok and Fail to better fit Rust
naming and be more distinguishable; and swap the value so that Ok is
zero and Fail is one.


Required for merge:
- [x] `pr-workflow-check / draft-release-check` is ✅ success
- Other rules GitHub shows you, or can be read in
[configuration](../terraform/github.com/branches.tf)

Makes review faster:
- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [x] Linked Zenhub/Github/Slack/etc reference if one exists
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [x] Added reviewer into `Reviewers`
- [ ] I tagged(`@`) or used other form of notification of one person who
I think can handle best review of this PR
- [x] I have proved that PR has no general regressions of relevant
features and processes required to release into production
- [x] Any dependency updates made, was done according guides from
relevant dependency
- Clicking all checkboxes
- Adding detailed description of changes when it feels appropriate (for
example when PR is big)