chore(price-feed): remove use of lazy-static
lazy-static is often no longer necessary now that OnceCell and OnceLock
are a thing.  Furthermore, in case of price-feed the lazy initialisation
isn’t even necessary: there’s so few entries in mapping that it’s enough
to use a static array than build a hash map.

Readjust the code to use said static array and remove dependency on
lazy-static.