feat(xcvm): use Displayed consistently for large integers (#4073)
For data structures which are part of public API consistently use
Dispalyed wrapper for u128 and u64 integers.  This is to make sure
that the numbers are encoded as strings when serialising to JSON.

This doesn’t change anything right now but future upgrades of
serde-json-wasm will change how integers are serialised so this is
future-proofing the code.

While at it, since Displayed is a shared wrapper used by various
modules, move it to shared module rather than keeping it in asset.


Required for merge:
- [ ] `pr-workflow-check / draft-release-check` is ✅ success
- Other rules GitHub shows you, or can be read in
[configuration](../terraform/github.com/branches.tf)

Makes review faster:
- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [x] Linked Zenhub/Github/Slack/etc reference if one exists
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [ ] Added reviewer into `Reviewers`
- [ ] I tagged(`@`) or used other form of notification of one person who
I think can handle best review of this PR
- [x] I have proved that PR has no general regressions of relevant
features and processes required to release into production
- [x] Any dependency updates made, was done according guides from
relevant dependency
- Clicking all checkboxes
- Adding detailed description of changes when it feels appropriate (for
example when PR is big)