feat(xcvm): instruction implementation and configuration with osmosis swap (#4083)
So actually implementation of swap instruction with config with mapping
for specific exchange of osmosis.

So need to thing about future designs in next iterations. First, is out
is always known from pool id, so out is just amount, not coin (but for
multi coint it is not the case). Second, if can have twap based slippage
as feature like in osmosis, instead of limit. Also if need to bool to
`suffice` like exchange like in XCM. Also if need exchange by delegation
(basically Transfer from by delegation or other way to use by
delegation).

Required for merge:
- [x] `pr-workflow-check / draft-release-check` is ✅ success
- Other rules GitHub shows you, or can be read in
[configuration](../terraform/github.com/branches.tf)

Makes review faster:
- [x] PR title is my best effort to provide summary of changes and has
clear text to be part of release notes
- [x] I marked PR by `misc` label if it should not be in release notes
- [x] Linked Zenhub/Github/Slack/etc reference if one exists
https://github.com/ComposableFi/composable/issues/3946
- [x] I was clear on what type of deployment required to release my
changes (node, runtime, contract, indexer, on chain operation, frontend,
infrastructure) if any in PR title or description
- [x] Added reviewer into `Reviewers`
- [x] I tagged(`@`) or used other form of notification of one person who
I think can handle best review of this PR
- [x] I have proved that PR has no general regressions of relevant
features and processes required to release into production
- [x] Any dependency updates made, was done according guides from
relevant dependency
- Clicking all checkboxes 
- Adding detailed description of changes when it feels appropriate (for
example when PR is big)