feat(xcvm): define helper proto::define_conversion macro (#4112)
To reduce amount of boilerplate when defining conversions between
protocol buffer messages and Rust types, introduce a define_conversion
macro.  It adds a `TryFrom<PB> for TY` and `From<TY> for PB` impls.

A bit of a downside of this is that it’s less obvious than an explicit
implementation of the traits.  However, with many of those traits
needed I think it’s worth it.