feat(xcvm): rework protocol messages [breaking change] (#4116)
Note that this is a breaking change.  Protobuf message definition as
well as serde definition of XCVM program and packets are changed.

1. Simplify protobuf definitions by removing wrapping types.  It’s not
protobuf style to have messages with just one flied (such as AssetId)
where instead the field can be used directly.  Removing those shortens
the protocol buffer and removes some of the conversion code.

2. Replace Self, Tip, Result and IpRegister BindingValues with
a single registry field and a corresponding Registry enum.  This
mirrors the way Rust types are defined and again simplifies conversion
code.

3. Rename `id` to `exchange_id` in instruction::Exchange and `network`
to `network_id` in instruction::Spawn.  This makes it more consistent
with protobuf messages as well as other types which use `foo_id` for
identifier fields.  This change affects serde.

4. Rename handful of fields in protobuf from camelCase to snake_case
as per the protobuf style guide.  This doesn’t change the Rust code.